+++ docs/conf.py
@@ -23,7 +23,7 @@ copyright = '2022, Federico Errica'
 author = 'Federico Errica'
 
 # The full version, including alpha/beta/rc tags
-release = '1.2.2'
+release = '1.2.3'
 
 
 # -- General configuration ---------------------------------------------------