+++ pydgn/data/splitter.py
@@ -112,7 +112,7 @@ class Splitter:
         self.outer_val_ratio = outer_val_ratio
         self.test_ratio = test_ratio
 
-    def get_graph_targets(self, dataset: pydgn.data.dataset.DatasetInterface) -> (bool, np.ndarray):
+    def get_targets(self, dataset: pydgn.data.dataset.DatasetInterface) -> (bool, np.ndarray):
         r"""
         Reads the entire dataset and returns the targets.
 
@@ -124,11 +124,16 @@ class Splitter:
             exception has not been thrown. The second value holds the actual targets or ``None``, depending on the
             first boolean value.
         """
-        try:
-            targets = np.array([d.y.item() for d in dataset])
+        if hasattr(dataset[0], 'y') and dataset[0].y is not None:
+            targets = torch.cat([d.y for d in dataset], dim=0).numpy()
             return True, targets
-        except Exception:
+        else:
             return False, None
+        # try:
+        #     targets = np.array([d.y.item() for d in dataset])
+        #     return True, targets
+        # except Exception:
+        #     return False, None
 
     @classmethod
     def load(cls, path: str):
@@ -326,7 +331,7 @@ class TemporalSplitter(Splitter):
         exception has not been thrown. The second value holds the actual targets or ``None``, depending on the
         first boolean value.
     """
-    def get_graph_targets(self, dataset: pydgn.data.dataset.TemporalDatasetInterface) -> Tuple[bool, np.ndarray]:
+    def get_targets(self, dataset: pydgn.data.dataset.TemporalDatasetInterface) -> Tuple[bool, np.ndarray]:
         raise NotImplementedError("You should subclass TemporalSplitter and implement this function!")
 
 
@@ -383,7 +388,7 @@ class SingleGraphSequenceSplitter(TemporalSplitter):
         assert not stratify, "You cannot stratify a single graph sequence."
 
 
-    def get_graph_targets(self, dataset: pydgn.data.dataset.TemporalDatasetInterface) -> Tuple[bool, np.ndarray]:
+    def get_targets(self, dataset: pydgn.data.dataset.TemporalDatasetInterface) -> Tuple[bool, np.ndarray]:
         try:
             targets = np.array([d.targets[-1].item() for d in dataset])
             return True, targets
@@ -449,6 +454,84 @@ class SingleGraphSequenceSplitter(TemporalSplitter):
                                val_idxs=inner_idxs[outer_val_idxs].tolist(),
                                test_idxs=outer_idxs[test_idxs].tolist())
         self.outer_folds.append(outer_fold)
+
+
+class SingleGraphSplitter(Splitter):
+    r"""
+    A splitter for a single graph dataset that randomly splits nodes into training/validation/test
+
+    Args:
+        n_outer_folds (int): number of outer folds (risk assessment). 1 means hold-out, >1 means k-fold
+        n_inner_folds (int): number of inner folds (model selection). 1 means hold-out, >1 means k-fold
+        seed (int): random seed for reproducibility (on the same machine)
+        stratify (bool): whether to apply stratification or not (should be true for classification tasks)
+        shuffle (bool): whether to apply shuffle or not
+        inner_val_ratio  (float): percentage of validation set for hold_out model selection. Default is ``0.1``
+        outer_val_ratio  (float): percentage of validation set for hold_out model assessment (final training runs). Default is ``0.1``
+        test_ratio  (float): percentage of test set for hold_out model assessment. Default is ``0.1``
+    """
+
+    def split(self, dataset: pydgn.data.dataset.DatasetInterface, targets: np.ndarray=None):
+        r"""
+        Compared with the superclass version, the only difference is that the range of indices spans across the number
+        of nodes of the single graph taken into consideration.
+        Args:
+            dataset (:class:`~pydgn.data.dataset.DatasetInterface`): the Dataset object
+            targets (np.ndarray]): targets used for stratification. Default is ``None``
+        """
+        assert len(dataset) == 1, "This class works only with single graph datasets"
+        idxs = range(dataset.data.x.shape[0])
+        stratified = self.stratify
+        outer_idxs = np.array(idxs)
+
+        # TODO all this code below could be reused from the original splitter
+        outer_splitter = self._get_splitter(
+            n_splits=self.n_outer_folds,
+            stratified=stratified,
+            eval_ratio=self.test_ratio)  # This is the true test (outer test)
+
+        for train_idxs, test_idxs in outer_splitter.split(outer_idxs, y=targets):
+
+            assert set(train_idxs) == set(outer_idxs[train_idxs])
+            assert set(test_idxs) == set(outer_idxs[test_idxs])
+
+            inner_fold_splits = []
+            inner_idxs = outer_idxs[train_idxs]  # equals train_idxs because outer_idxs was ordered
+            inner_targets = targets[train_idxs] if targets is not None else None
+
+            inner_splitter = self._get_splitter(
+                n_splits=self.n_inner_folds,
+                stratified=stratified,
+                eval_ratio=self.inner_val_ratio)  # The inner "test" is, instead, the validation set
+
+            for inner_train_idxs, inner_val_idxs in inner_splitter.split(inner_idxs, y=inner_targets):
+                inner_fold = InnerFold(train_idxs=inner_idxs[inner_train_idxs].tolist(),
+                                       val_idxs=inner_idxs[inner_val_idxs].tolist())
+                inner_fold_splits.append(inner_fold)
+
+                # False if empty
+                assert not bool(set(inner_train_idxs) & set(inner_val_idxs) & set(test_idxs))
+                assert not bool(set(inner_idxs[inner_train_idxs]) & set(inner_idxs[inner_val_idxs]) & set(test_idxs))
+
+            self.inner_folds.append(inner_fold_splits)
+
+            # Obtain outer val from outer train in an holdout fashion
+            outer_val_splitter = self._get_splitter(n_splits=1, stratified=stratified,
+                                                    eval_ratio=self.outer_val_ratio)
+            outer_train_idxs, outer_val_idxs = list(outer_val_splitter.split(inner_idxs, y=inner_targets))[0]
+
+            # False if empty
+            assert not bool(set(outer_train_idxs) & set(outer_val_idxs) & set(test_idxs))
+            assert not bool(set(outer_train_idxs) & set(outer_val_idxs) & set(test_idxs))
+            assert not bool(set(inner_idxs[outer_train_idxs]) & set(inner_idxs[outer_val_idxs]) & set(test_idxs))
+
+            np.random.shuffle(outer_train_idxs)
+            np.random.shuffle(outer_val_idxs)
+            np.random.shuffle(test_idxs)
+            outer_fold = OuterFold(train_idxs=inner_idxs[outer_train_idxs].tolist(),
+                                   val_idxs=inner_idxs[outer_val_idxs].tolist(),
+                                   test_idxs=outer_idxs[test_idxs].tolist())
+            self.outer_folds.append(outer_fold)
 
 
 class LinkPredictionSingleGraphSplitter(Splitter):
@@ -693,3 +776,4 @@ class LinkPredictionSingleGraphSplitter(Splitter):
         splitter_args.update({"undirected": self.undirected,
                               "avoid_opposite_negative_edges": self.avoid_opposite_negative_edges})
         return splitter_args
+